refactor(discord): move Discord UI from core/ to adapters/discord/ui (client side)
