+++ mypalclara/adapters/discord/main.py
@@ -35,7 +35,7 @@ from mypalclara.adapters.discord.channel_modes import get_channel_mode
 from mypalclara.adapters.discord.gateway_client import DiscordGatewayClient
 from mypalclara.adapters.discord.voice import VoiceManager
 from mypalclara.config.logging import get_logger, init_logging
-from mypalclara.core.discord import setup as setup_slash_commands
+from mypalclara.adapters.discord.ui import setup as setup_slash_commands
 
 init_logging()
 logger = get_logger("adapters.discord")
