+++ mypalclara/adapters/discord/ui/embeds.py
